requirements.txt
# Vāṇī-Vision Requirements
# Core Framework
streamlit>=1.32.0

# OCR Engine
paddleocr>=2.7.3
paddlepaddle>=2.6.0   # CPU-only build

# LLM Inference (CPU-optimized quantized models)
llama-cpp-python>=0.2.57

# Image Processing
opencv-python>=4.9.0.80
Pillow>=10.2.0
numpy>=1.26.4

# Multilingual / NLP utilities
langdetect>=1.0.9
deep-translator>=1.11.4  # offline-capable fallback translation

# UI Extras
streamlit-extras>=0.4.2
plotly>=5.20.0

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.2
